Fixed precision bug in distance transform for nfp wih dimensions > 4096 (temporary removed image output though).

+++ RasterVoronoiPreProcessor/dt/dt.h
@@ -28,16 +28,16 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307 USA
 #define INF 1E20f
 
 /* dt of 1d function using squared distance */
-static float *dt(float *f, int n) {
-  float *d = new float[n];
+static double *dt(double *f, int n) {
+  double *d = new double[n];
   int *v = new int[n];
-  float *z = new float[n+1];
+  double *z = new double[n+1];
   int k = 0;
   v[0] = 0;
   z[0] = -INF;
   z[1] = +INF;
   for (int q = 1; q <= n-1; q++) {
-    float s  = ((f[q]+square(q))-(f[v[k]]+square(v[k])))/(2*q-2*v[k]);
+    double s  = ((f[q]+square(q))-(f[v[k]]+square(v[k])))/(2*q-2*v[k]);
     while (s <= z[k]) {
       k--;
       s  = ((f[q]+square(q))-(f[v[k]]+square(v[k])))/(2*q-2*v[k]);
@@ -61,17 +61,17 @@ static float *dt(float *f, int n) {
 }
 
 /* dt of 2d function using squared distance */
-static void dt(image<float> *im) {
+static void dt(image<double> *im) {
   int width = im->width();
   int height = im->height();
-  float *f = new float[std::max(width,height)];
+  double *f = new double[std::max(width,height)];
 
   // transform along columns
   for (int x = 0; x < width; x++) {
     for (int y = 0; y < height; y++) {
       f[y] = imRef(im, x, y);
     }
-    float *d = dt(f, height);
+    double *d = dt(f, height);
     for (int y = 0; y < height; y++) {
       imRef(im, x, y) = d[y];
     }
@@ -83,7 +83,7 @@ static void dt(image<float> *im) {
     for (int x = 0; x < width; x++) {
       f[x] = imRef(im, x, y);
     }
-    float *d = dt(f, width);
+    double *d = dt(f, width);
     for (int x = 0; x < width; x++) {
       imRef(im, x, y) = d[x];
     }
@@ -95,11 +95,11 @@ static void dt(image<float> *im) {
 
 
 /* dt of binary image using squared distance */
-static image<float> *dt(image<uchar> *im, uchar on = 1) {
+static image<double> *dt(image<uchar> *im, uchar on = 1) {
   int width = im->width();
   int height = im->height();
 
-  image<float> *out = new image<float>(width, height, false);
+  image<double> *out = new image<double>(width, height, false);
   for (int y = 0; y < height; y++) {
     for (int x = 0; x < width; x++) {
       if (imRef(im, x, y) == on)
